+++ server/prisma-rs/query-engine/core/src/executor/mod.rs
@@ -1,8 +1,13 @@
 //! A slightly more generic interface over executing read and write queries
 #![warn(warnings)]
 
+mod read;
+mod write;
+
+pub use read::ReadQueryExecutor;
+pub use write::WriteQueryExecutor;
+
 use crate::{Query, WriteQuery, WriteQueryResult, ReadQuery, ReadQueryResult};
-use crate::{WriteQueryExecutor, ReadQueryExecutor};
 
 /// A wrapper around QueryExecutor
 pub struct Executor {
@@ -14,11 +19,9 @@ impl Executor {
 
     /// Can be given a list of both ReadQueries and WriteQueries
     ///
-    /// Will execute WriteQueries first, then all ReadQueries, while preserving order
+    /// Will execute WriteQueries first, then all ReadQueries, while preserving order.
     pub fn exec_all(&mut self, queries: Vec<Query>) {
         let (writes, reads) = Self::split_read_write(queries);
-
-
     }
 
     fn split_read_write(queries: Vec<Query>) -> (Vec<WriteQuery>, Vec<Option<ReadQuery>>) {
